+++ src/onPreprocess.js
@@ -1,4 +1,4 @@
-import { updateXDomain } from './util/updateXDomain';
+import updateXDomain from './onPreprocess/updateXDomain';
 
 export default function onPreprocess() {
     const chart = this,
